fixed issue with git file renaming
