+++ src/features/app/App.jsx
@@ -15,8 +15,8 @@ import Team from '../team/Team';
 import Signin from '../user/Signin';
 import Signup from '../user/Signup';
 import SignedInUser from '../user/SignedInUser';
-import LandingPage from '../landingpage/LandingPage';
-import DashBoard from '../dashboard/DashBoard';
+import LandingPage from '../landingpage/LandingPage2';
+import DashBoard from '../dashboard/DashBoard2';
 
 function FallBack() {
   return (
